have done component for the form, created Provider for all app
